tools/proof_factory/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proof-factory"
version = "0.1.0"
description = "Proof Closure Factory for Lean obligations in goldbach-horizon"
requires-python = ">=3.11"
dependencies = ["pyyaml>=6.0", "mpmath>=1.3.0"]

[tool.setuptools]
packages = ["pf"]
